fix: address 5-persona review findings
- Extract dlog_m_to_linear/linear_to_dlog_m to pipeline_utils.py
  (eliminates duplication between 00_generate_lut.py and
  generate_references.py)
- Fix correct_image() defaults to reference DEFAULT_PARAMS dict
  (was using stale values: 1.35/1.05/0.95 vs correct 2.0/1.0/0.75)
- Add highlight rolloff to generate_references.py to match
  00_generate_lut.py correction chain
- Add error handling for corrupt/unreadable images in both scripts
- Regenerate references and LUT with fixed code
